# backend/runtime/src/main/java/com/jabiz/runtime/web/EntityController.java
package com.jabiz.runtime.web;

import com.jabiz.dataset.DatasetDefinition;
import com.jabiz.entity.EntityDefinition;
import com.jabiz.entity.ValidationException;
import com.jabiz.entity.Violation;
import com.jabiz.i18n.PlatformErrorCodes;
import com.jabiz.query.EntityQuery;
import com.jabiz.runtime.DatasetEntityManager;
import com.jabiz.runtime.EntityInstance;
import com.jabiz.runtime.EntityNotFoundException;
import com.jabiz.runtime.dataset.DatasetRegistry;
import com.jabiz.runtime.entity.EntityDefinitionRegistry;
import com.jabiz.runtime.process.ProcessExecutor;
import com.jabiz.runtime.process.entity.AddEntityInput;
import com.jabiz.runtime.process.entity.AddProcessDefinition;
import com.jabiz.runtime.process.entity.DeleteEntityInput;
import com.jabiz.runtime.process.entity.DeleteProcessDefinition;
import com.jabiz.runtime.process.entity.UpdateEntityInput;
import com.jabiz.runtime.process.entity.UpdateProcessDefinition;
import com.jabiz.process.ProcessDefinition;
import com.jabiz.runtime.context.RequestContexts;
import com.jabiz.runtime.security.MfaPolicy;
import com.jabiz.runtime.security.Permissions;
import com.jabiz.runtime.security.SensitiveDataMasker;
import org.springframework.core.env.Environment;
import org.springframework.core.env.Profiles;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;

import java.util.List;
import java.util.Map;

/**
 * Entity-independent HTTP API. Writes run the generic add, update and delete processes; the
 * list endpoint reads through the dataset that serves the entity type.
 *
 * Every entity type that has a dataset and a version field is writable through this API. Reads need the read
 * permission of the entity's default dataset; writes its write permission and the permissions of the generic process.
 */
@RestController
@RequestMapping("/api/entities/{entityType}")
class EntityController {

    /** Body of a partial update: the version the caller last read and the fields to change. */
    record UpdateRequest(Long version, Map<String, Object> attributes) {}

    private final ProcessExecutor processes;
    private final EntityDefinitionRegistry entities;
    private final DatasetRegistry datasets;
    private final DatasetEntityManager entityManager;
    private final SensitiveDataMasker masker;
    private final boolean development;
    private final MfaPolicy mfa;

    EntityController(
        ProcessExecutor processes,
        EntityDefinitionRegistry entities,
        DatasetRegistry datasets,
        DatasetEntityManager entityManager,
        SensitiveDataMasker masker,
        Environment environment,
        MfaPolicy mfa
    ) {
        this.mfa = mfa;
        this.processes = processes;
        this.entities = entities;
        this.datasets = datasets;
        this.entityManager = entityManager;
        this.masker = masker;
        this.development = environment.acceptsProfiles(Profiles.of("dev"));
    }

    /** Lists instances; {@code sort} is a field name, prefixed with '-' for descending order. */
    @GetMapping
    Flux<EntityInstance> list(
        @PathVariable String entityType,
        @RequestParam(defaultValue = "50") int limit,
        @RequestParam(defaultValue = "0") int offset,
        @RequestParam(required = false) String sort
    ) {
        return RequestContexts.current().flatMapMany(context -> {
            EntityDefinition definition = entities.find(entityType).orElseThrow(
                () -> new EntityNotFoundException("Unregistered entity type: " + entityType));
            DatasetDefinition dataset = defaultDataset(entityType);
            Permissions.requireDeclared(context, dataset.permissions().read(), development,
                "Reading " + entityType + " through dataset " + dataset.resourceId());
            if (limit <= 0) {
                throw invalid("limit", "limit must be positive");
            }
            if (offset < 0) {
                throw invalid("offset", "offset must not be negative");
            }

            EntityQuery.Builder query = EntityQuery.builder().limit(limit).offset(offset);
            if (sort != null && !sort.isBlank()) {
                boolean ascending = !sort.startsWith("-");
                String field = ascending ? sort : sort.substring(1);
                if (definition.findField(field).isEmpty()) {
                    throw invalid("sort", "Entity '" + entityType + "' has no field '" + field + "'");
                }
                if (definition.field(field).sensitive()) {
                    // Even the order of secrets tells something about them (docs/design/10-security.md section 6).
                    throw new ValidationException(List.of(new Violation(field, PlatformErrorCodes.SORT_NOT_ALLOWED,
                        "Sorting by [" + field + "] is not allowed")));
                }
                query.orderBy(field, ascending);
            }
            return entityManager.query(dataset, definition, query.build()).map(masker::hide);
        });
    }

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    Mono<EntityInstance> add(@PathVariable String entityType, @RequestBody Map<String, Object> attributes) {
        return write(entityType, AddProcessDefinition.DEFINITION)
            .then(Mono.defer(() -> processes.execute(
                AddProcessDefinition.DEFINITION, new AddEntityInput(entityType, attributes))))
            .map(masker::hide);
    }

    @PatchMapping("/{id}")
    Mono<EntityInstance> update(
        @PathVariable String entityType, @PathVariable String id, @RequestBody UpdateRequest request
    ) {
        return write(entityType, UpdateProcessDefinition.DEFINITION)
            .then(Mono.defer(() -> processes.execute(
                UpdateProcessDefinition.DEFINITION,
                new UpdateEntityInput(entityType, id, request.version(), request.attributes()))))
            .map(masker::hide);
    }

    @DeleteMapping("/{id}")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    Mono<Void> delete(
        @PathVariable String entityType, @PathVariable String id, @RequestParam(required = false) Long version
    ) {
        return write(entityType, DeleteProcessDefinition.DEFINITION)
            .then(Mono.defer(() -> processes.execute(
                DeleteProcessDefinition.DEFINITION, new DeleteEntityInput(entityType, id, version))))
            .then();
    }

    /**
     * Writes run a generic process through the entity's default dataset: the caller needs the process's permissions
     * and the dataset's write permission.
     */
    private Mono<Void> write(String entityType, ProcessDefinition<?, ?, ?> process) {
        return RequestContexts.current().doOnNext(context -> {
            Permissions.requireAll(context, process.permissions(), development, "Process " + process.name());
            DatasetDefinition dataset = defaultDataset(entityType);
            Permissions.requireDeclared(context, dataset.permissions().write(), development,
                "Writing " + entityType + " through dataset " + dataset.resourceId());
            mfa.require(context, dataset.policy().writeMfa(), "Writing " + entityType);
        }).then();
    }

    private DatasetDefinition defaultDataset(String entityType) {
        return datasets.findForEntity(entityType).orElseThrow(
            () -> new EntityNotFoundException("No dataset serves entity type: " + entityType));
    }

    private static ValidationException invalid(String field, String message) {
        return new ValidationException(List.of(new Violation(field, "INVALID_VALUE", message)));
    }
}