# frontend/openapi/openapi.json
{
  "components" : {
    "schemas" : {
...
      "EntityInstance" : {
        "properties" : {
          "attributes" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "entityType" : {
            "type" : "string"
          },
          "id" : { },
          "state" : {
            "type" : "string"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "UpdateRequest" : {
        "properties" : {
          "attributes" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/entities/{entityType}" : {
      "get" : {
        "operationId" : "list",
        "parameters" : [ {
          "in" : "path",
          "name" : "entityType",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 50,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "offset",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/EntityInstance"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "entity-controller" ]
      },
      "post" : {
        "operationId" : "add",
        "parameters" : [ {
          "in" : "path",
          "name" : "entityType",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "additionalProperties" : { },
                "type" : "object"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityInstance"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "entity-controller" ]
      }
    },
    "/api/entities/{entityType}/{id}" : {
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "in" : "path",
          "name" : "entityType",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "version",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "entity-controller" ]
      },
      "patch" : {
        "operationId" : "update",
        "parameters" : [ {
          "in" : "path",
          "name" : "entityType",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityInstance"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "entity-controller" ]
      }
    },
...
  }
}